Does things (#47)
* pinsir icon_state fixed

glaceon north sprite also shifted left 1 pixel

* stuff

* moves to global lists so you can have more than 1 pod

* oops

* gender selection

* fixes ninetales sprite

* Update pokemon.dm

* ghost verb for mon spawning

+++ content_arfs/code/ghost_pods_arfs.dm
@@ -1,3 +1,7 @@
+//globals so the lists dont get generated more than once
+var/global/list/pokemon_choices_list = list()//Referenced list for use in input()
+var/global/list/pokemon_pods = list()//List of pods that ghosts can spawn at
+
 //Pod to spawn in as pokemon or other mobs.
 /obj/structure/ghost_pod/ghost_activated/pokemon
 	name = "\improper Pokemon resleever"
@@ -8,46 +12,58 @@
 	icon_state_opened = "borg_pod_opened"
 	anchored = TRUE
 	var/do_announcement = TRUE //FALSE won't give them a choice of announcing
-	var/list/p_list = list()
-	var/list/p_list_types = /mob/living/simple_mob/animal/passive/pokemon //Subtypes of this will be added to p_list
-	var/list/p_list_paths = list()
-	var/list/remove_paths = list(/mob/living/simple_mob/animal/passive/pokemon/leg, //These are removed from the final list of types
-						    /mob/living/simple_mob/animal/passive/pokemon,
+	var/list/remove_paths = list(/mob/living/simple_mob/animal/passive/pokemon/leg, 		//Can tailor this list for different types of
+						    /mob/living/simple_mob/animal/passive/pokemon,					//spawners down the line
 						    /mob/living/simple_mob/animal/passive/pokemon/eevee/jolteon/bud)
 
-//Should move this into a global list if we need it for anything else.
 /obj/structure/ghost_pod/ghost_activated/pokemon/Initialize()
 	. = ..()
-	p_list_paths = typesof(/mob/living/simple_mob/animal/passive/pokemon) - remove_paths//Create a list of mob paths
-	for (var/path in p_list_paths)//add the mobs to a list with their names referencing paths
+	generate_lists()
+	var/turf/T = get_turf(src)
+	pokemon_pods["[T.loc]"] = src
+
+/obj/structure/ghost_pod/ghost_activated/pokemon/proc/generate_lists()
+	if(LAZYLEN(pokemon_choices_list))
+		return FALSE	//The global list is already generated
+	var/pokemon_choices_list_paths = typesof(/mob/living/simple_mob/animal/passive/pokemon) - remove_paths
+	for (var/path in pokemon_choices_list_paths)//add the mobs to a list with their names referencing paths
 		var/mob/living/simple_mob/animal/passive/pokemon/P = new path()
-		p_list["[P.name]"] = P.type
+		pokemon_choices_list["[P.name]"] = P.type
 		qdel(P)
+
+/obj/structure/ghost_pod/manual/attack_hand(var/mob/living/user)
+	//Inform curious minds about how to use this.
+	to_chat(user, "<span class='notice'>You can't seem to find a way to interact with this from the outside.</span>")
+	to_chat(user, "<span class='warning'>(Only observers may use [src])</span>")
+	return
 
 /obj/structure/ghost_pod/ghost_activated/pokemon/attack_ghost(var/mob/observer/dead/user)
 	if (ticker.current_state != GAME_STATE_PLAYING)
 		to_chat(user, "<span class='warning'>The round either hasn't started yet or has ended.</span>")
 		return
-	if (p_list == list() || !p_list)
-		to_chat(user, "<span class='warning'>Pod configuration error.</span>")
+	if (!LAZYLEN(pokemon_choices_list))
+		to_chat(user, "<span class='warning'>Pod configuration error. Report this to a developer.</span>")
 		return
 	create_occupant(user)
-
 
 /obj/structure/ghost_pod/ghost_activated/pokemon/create_occupant(var/mob/M)
 	var/m_ckey = M.ckey
 	var/turf/T
 	var/area/A
-	var/p_choice = input(M, "What would you like to spawn in as?", "[src.name]") as null|anything in p_list
+	var/p_choice = input(M, "What would you like to spawn in as?", "[src.name]") as null|anything in pokemon_choices_list
 	if(!p_choice || isnull(p_choice))
 		to_chat(M, "<span class='notice'>Spawning aborted.</span>")
 		return
-	p_choice = p_list["[p_choice]"]
+	p_choice = pokemon_choices_list["[p_choice]"]
 	var/newname = input(M, "Would you like to change your name or use the default one? Enter nothing to use the default name. Canceling will stop the spawning process.", "Name Change") as null|text
 	if(isnull(newname))
 		to_chat(M, "<span class='notice'>Spawning aborted.</span>")
 		return
 	newname = sanitize(newname, MAX_NAME_LEN)//Sanitize the name afterwards, so we know if they hit cancel or input an empty string
+	var/new_gender = input(M, "Choose your Pokemon's gender:", "Character Preference", "neuter") as null|anything in list("neuter", "male", "female")
+	if(isnull(new_gender))
+		to_chat(M, "<span class='notice'>Spawning aborted.</span>")
+		return
 	var/announce_choice = FALSE
 	if(do_announcement)
 		announce_choice = input(M, "Would you like to announce your arrival over the common radio channel?", "[src.name]") as null|anything in list("Yes","No")
@@ -59,10 +75,12 @@
 	T = get_turf(src)
 	A = T.loc
 	var/mob/living/simple_mob/animal/passive/pokemon/P = new p_choice(T)
-	if(newname)
+	if(newname)//Still not empty after sanitization
 		P.name = newname
 		P.voice_name = P.name
 	P.real_name = P.name
+	if(new_gender)
+		P.gender = new_gender
 	if(M.mind)
 		M.mind.transfer_to(P)
 	if(m_ckey)
